Extract inner GrpcDirectServerStreamingCallable class to toplevel
